chore(scripts): remove scripts and move path
