+++ src/router/index.js
@@ -1,9 +1,17 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import { collectRoute } from '../../scripts/collectRouter'
-// const path = require.context('../../docs/zh-cn/', true, /.md$/)
 const path = require.context('../../docs/zh-cn', true, /.md$/)
 
-collectRoute(path)
+const collectRoute = (context) => {
+  return context.keys().map((p) => {
+    const sourceName = p.match(/\w+(?=.md)/g).toString()
+    const routeName = sourceName.charAt(0).toUpperCase() + sourceName.slice(1)
+    return {
+      path: `/${routeName}`,
+      name: routeName,
+      component: context(p).default,
+    }
+  })
+}
 
 const routes = [
   { path: '/', redirect: { name: 'Introduce' } },
